Allow only measuring SPMV performance in CGSOLVE

+++ papers/IWOCL2024/cgsolve/generate_matrix.hpp
@@ -24,7 +24,7 @@
 
 #include<Kokkos_Core.hpp>
 
-using INT_TYPE = int;
+using INT_TYPE = int64_t;
 
 template<class MemSpace>
 struct CrsMatrix {
